LPD-63086 Rename CopyFragmentModal to FragmentSetModal
