+++ modules/apps/fragment/fragment-web/src/main/resources/META-INF/resources/js/ViewContributedFragmentEntriesManagementToolbarPropsTransformer.js
@@ -4,7 +4,7 @@
  */
 
 import {render} from '@liferay/frontend-js-react-web';
-import {CopyFragmentModal} from '@liferay/layout-js-components-web';
+import {FragmentSetModal} from '@liferay/layout-js-components-web';
 import {getCheckedCheckboxes} from 'frontend-js-web';
 
 export default function propsTransformer({
@@ -29,7 +29,7 @@ export default function propsTransformer({
 		);
 
 		render(
-			CopyFragmentModal,
+			FragmentSetModal,
 			{
 				addFragmentCollectionURL,
 				contributedEntryKeys: contributedEntryKeys.split(','),
