+++ modules/apps/layout/layout-js-components-web/src/main/resources/META-INF/resources/js/index.ts
@@ -5,7 +5,7 @@
 
 export {default as InstallFragmentModal} from '././components/marketplace/InstallFragmentModal';
 export {default as CardStyleModal} from '././components/modals/CardStyleModal';
-export {default as CopyFragmentModal} from '././components/modals/CopyFragmentModal';
+export {default as FragmentSetModal} from '././components/modals/FragmentSetModal';
 export {
 	default as ColorPicker,
 	DEFAULT_TOKEN_LABEL,
