+++ modules/apps/layout/layout-js-components-web/test/components/modals/FragmentSetModal.test.js
@@ -8,17 +8,17 @@ import {act, fireEvent, render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import CopyFragmentModal from '../../../src/main/resources/META-INF/resources/js/components/modals/CopyFragmentModal';
+import FragmentSetModal from '../../../src/main/resources/META-INF/resources/js/components/modals/FragmentSetModal';
 
 const renderComponent = ({fragmentCollections = []} = {}) => {
 	const user = userEvent.setup({advanceTimers: jest.advanceTimersByTime});
 
-	render(<CopyFragmentModal fragmentCollections={fragmentCollections} />);
+	render(<FragmentSetModal fragmentCollections={fragmentCollections} />);
 
 	return user;
 };
 
-describe('CopyFragmentModal', () => {
+describe('FragmentSetModal', () => {
 	beforeAll(() => {
 		jest.useFakeTimers();
 	});
